pyproject.toml
[tool.poetry]
name = "scrybot"
version = "0.1.0"
description = "A scryfall discord bot"
authors = ["Adam Weidner"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
discord-py = "^2.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
